django3/mysite/xbee/requirements.txt
colored==1.4.2
digi-xbee==1.3.0
paho-mqtt==1.5.1
pyserial==3.5
six==1.15.0
SQLAlchemy==1.3.22
srp==1.0.16
